evals refactor
