+++ evals/helpers.py
@@ -3,6 +3,8 @@ import os
 import mlflow
 from dotenv import load_dotenv
 from mlflow.openai import autolog
+import json
+from typing import List, TypedDict
 
 load_dotenv()
 
@@ -18,3 +20,24 @@ def setup_mlflow(experiment_name: str):
         mlflow.set_tracking_uri("databricks://memorymodule-evals")
         mlflow.set_experiment(experiment_name)
     autolog()
+
+class SessionMessage(TypedDict):
+    content: str
+    role: str
+
+class DatasetItem(TypedDict):
+    category: str
+    session: List[SessionMessage]
+    query: str
+    expected_strings_in_memories: List[str]
+
+class Dataset(TypedDict):
+    version: int
+    title: str
+    description: str
+    data: List[DatasetItem]
+
+def load_dataset() -> Dataset:
+    with open("evals/memory_module_dataset.json") as f:
+        dataset = json.load(f)
+        return dataset
